another round of refactoring.
